Ajout des dossiers back-add

+++ app/Http/Controllers/UserController.php
@@ -57,12 +57,11 @@ public function edit(User $user){
     }}
 
     public function destroy(User $user){
-        $user->delete();
-
-        if(request()->wantsJson()){
-            return response()->json([
-                'message'=>'User deleted successfully'
-            ]);
+        if (!$user) {
+            return response()->json(['message' => 'User not found'], 404);
         }
+    
+        $user->delete();
+        return response()->json(['message' => 'User deleted successfully'], 200);
     }
 }
